Optical components

Optical elements serve as the building blocks for optical systems in the context of this package, representing components such as mirrors, lenses, filters and so on. A collection of basic optical elements is provided with this package as is. They are tested for the correctness of their optical interactions and are verified to work with reasonable fidelity. For detailed documentation, refer to the following table of contents.

Component overview

Mirrors

Plano Mirrors

Concave Mirrors

Lenses

Constructing lens shapes

Spherical lenses

Aspherical lenses

Cylindrical lenses

Doublet lenses

Beamsplitters

Thin beamsplitter

Plate beamsplitters

Cube beamsplitter

Detectors

Photodetector type

Spotdetector type

Point-spread-function detector type

Polarizers

Jones matrix element representation

Polarisation filter

Moving optical elements

Optical elements can move around freely in three-dimensional space, which enables the modeling of kinematics within optical setups. When objects are manipulated, they are translated and rotated around their self-defined center of gravity, which is represented as a $\mathbb{R}^3$-vector and will be referred to as its position. Additionally, the orientation of an object, defined as its local fixed coordinate system, is represented by an orthonormal matrix in $\mathbb{R}^3$. If the object is rotated, this matrix can be used to calculate the inverse transform into global coordinates.

Optical system kinematics

Elements can be moved freely between each call of solve_system!. However, during tracing it is assumed that all elements remain static.

For elements that implement the BeamletOptics.AbstractObject interface, the following movement commands are provided:

Translation

translate3d!

translate_to3d!

Rotation

rotate3d!

xrotate3d!

yrotate3d!

zrotate3d!

align3d!

Reset commands

reset_translation3d!

reset_rotation3d!

Relative motion

Unless specified otherwise, the translation and rotation commands result in relative motions to the current position and orientation. This must be taken into account when trying to model a specific set of movements.

Groups of optical elements

For the easier representation of a group of BeamletOptics.AbstractObjects that moves as one, the ObjectGroup can be used. Refer to the Lens groups example for more information.

BeamletOptics.ObjectGroup — Type

ObjectGroup <: AbstractObjectGroup

A tree-like storage container for groups of objects. Can store individual objects and subgroups. Main purpose is handling of, i.e., groups of lenses.

Fields

center: a point in 3D space which is regarded as the reference origin of the group

dir: a 3x3 matrix that describes the common orientation of the group

objects: stores AbstractObject, can also store subgroups of type AbstractObjectGroup

Kinematic

A ObjectGroup implements the kinematic functions of AbstractObject. The following logic is applied to

translate3d!: all objects in the group are translated by the offset vector

translate_to3d!: all objects are moved in parallel such that the group center is equal to the target position

rotate3d!: all objects are rotated around the center point with respect to their relative position

Listing available components

When using this package in the REPL, a tree view of all implemented BeamletOptics.AbstractObjects can be generated via the BeamletOptics.list_subtypes helper function. Note that this function is not able to determine all available constructors.

julia> BeamletOptics.list_subtypes(BeamletOptics.AbstractObject);└── BeamletOptics.AbstractObject
    ├── BeamletOptics.AbstractBeamsplitter{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   ├── BeamletOptics.AbstractPlateBeamsplitter{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   │   ├── RectangularPlateBeamsplitter{T} where T<:Real
    │   │   └── RoundPlateBeamsplitter{T} where T<:Real
    │   ├── CubeBeamsplitter{T} where T<:Real
    │   └── ThinBeamsplitter
    ├── BeamletOptics.AbstractDetector{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   ├── PSFDetector{T} where T<:Real
    │   ├── Photodetector{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   └── Spotdetector{T} where T<:Real
    ├── BeamletOptics.AbstractJonesPolarizer{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   └── PolarizationFilter{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    ├── BeamletOptics.AbstractObjectGroup{T} where T<:Real
    │   └── ObjectGroup{T} where T<:Real
    ├── BeamletOptics.AbstractReflectiveOptic{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   ├── ConcaveSphericalMirror{T} where T<:Real
    │   ├── Mirror{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   ├── Retroreflector{T} where T<:Real
    │   ├── RightAnglePrismMirror{T} where T<:Real
    │   └── RoundPlanoMirror{T} where T<:Real
    ├── BeamletOptics.AbstractRefractiveOptic{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   ├── BeamletOptics.AbstractDoubletRefractiveOptic{T, F, B} where {T<:Real, F<:BeamletOptics.AbstractShape{T}, B<:BeamletOptics.AbstractShape{T}}
    │   │   └── DoubletLens{T, F, B} where {T<:Real, F<:BeamletOptics.AbstractShape{T}, B<:BeamletOptics.AbstractShape{T}}
    │   ├── Lens{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    │   └── Prism{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    ├── IntersectableObject{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}
    └── NonInteractableObject{T, S} where {T<:Real, S<:BeamletOptics.AbstractShape{T}}

At least 27 types have been found.